Adds standard deviation to position and velocity stats
Adds standard deviation calculations to the position (lat/lon/alt) and
velocity (ENU) statistics displayed in the UI.

Displays standard deviation alongside the mean for each of these values.
Corrects the calculation of deviation for latitude and longitude to
account for the curvature of the earth, and displays standard
deviation in meters.

Changes histogram generation for lat/lon to use pre-computed
deviations, to avoid recalculating them.

Adds labels to the histogram x-axes to indicate the deviation units.